pyproject.toml
[tool.poetry]
name = "project-practice"
version = "0.1.0"
description = ""
authors = ["jiwon hwang <realjiwon98@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
